+++ CHANGELOG.md
@@ -35,6 +35,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Avoid eager link type creation for relation links, only create and fetch them on first render.
 
 #### 💅 Polish
+- **[💥Breaking]** Use typed `TemplateState` for `Element.elementState` and `Link.linkState` to avoid accidental type mismatch.
 - Make dialogs fill the available viewport when the viewport width is small:
   * This is controlled by new CSS property `--reactodia-dialog-viewport-breakpoint-s` with default value `600px` which makes dialog fill the viewport if the available width is less or equal to that value.
 - Allow to override base z-index level for workspace components with a set z-index value via `--reactodia-z-index-base` CSS property;
@@ -56,9 +57,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
   * Change CSS class for default link template from `reactodia-default-link` to `reactodia-standard-link`;
   * Change translation groups from `standard_template` / `default_link_template` to `standard_element` / `standard_link`.
 - Move "expand/collapse on double click" global element behavior to `StandardEntity` and `ClassicEntity` implementation only.
-- Make template state properties typed (only affects TypeScript types, the representation is the same):
-  * Use `templateProperty('key').of<Type>()` to define template properties;
-  * Use `getTemplateProperty()` and `setTemplateProperty()` functions to get or set/unset template state properties.
 - Change `MetadataProvider.{createEntity, createRelation}` to return result object with initial template state in addition to the data to customize the created cells (i.e. new elements can be expanded or collapsed).
 - Add `EditorController.applyAuthoringChanges()` method to apply current authoring changes to the diagram (i.e. change entity data, delete relations, etc) and reset the change state to be empty.
 - Deprecate and hide by default "Edit" and "Delete" action buttons in `StandardEntity` expanded state (can be re-enabled by setting `showActions` prop to `true`).